+++ src/main/java/com/core/arnuv/repository/ITarifarioRapository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.core.arnuv.model.Tarifario;
 
 @Repository
-public interface ITarifarioRapository extends JpaRepository<Tarifario, Long> {
+public interface ITarifarioRapository extends JpaRepository<Tarifario, Integer> {
 
 }
